drivers: clock system is working and running at full speed
Might wanna take a look a the code again. The ROM code
does some initialization dureing boot, but does not
reset the clock network. Therefore the main clock and
master clock source need to change before configuring
the system. Currently the master clock is reset to the
main clock in the main clock select and the PLLA config.
This however is not ideal!